Separated StashKey and Stockpile
