+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -112,7 +112,9 @@
     <Compile Include="Objects.cs" />
     <Compile Include="Render.cs" />
     <Compile Include="BlastDiff.cs" />
+    <Compile Include="StashKey.cs" />
     <Compile Include="StepActions.cs" />
+    <Compile Include="Stockpile.cs" />
     <Compile Include="StockpileManager_UISide.cs" />
     <Compile Include="StockpileManager_EmuSide.cs" />
   </ItemGroup>
